rename nginx parser to server_log parser
